Extract Simon demo messages to dedicated module mirroring templates pattern #123
