+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@joshuafolkken/game-kit",
-	"version": "0.46.0",
+	"version": "0.47.0",
 	"keywords": [
 		"svelte"
 	],
